Fixes for #60
Support the 'runtime' configuration of the domain name for the system by changing locations where the hard-coded default domain name is apparent. Mostly it's in adding Jinja2 `raw` flags around variables so that the substitution can happen using the variable created in `ipconf.sh`.

Note that some adjustment of the `regen-certificates.yml` playbook may be needed due to #62 (still investigating).